+++ src/Contract/ExceptionInterface.php
@@ -6,5 +6,5 @@ namespace Zjk\SqlTwig\Contract;
 
 interface ExceptionInterface
 {
-    public static function formatted(string $message, string $exceptionMessage, bool $isDebug): self;
+    public static function formatted(string $message, string $exceptionMessage, bool $isDebug, ?\Throwable $throwable = null): self;
 }