+++ src/Exception/Message/ExceptionMessage.php
@@ -6,7 +6,7 @@ namespace Zjk\SqlTwig\Exception\Message;
 
 trait ExceptionMessage
 {
-    public static function formatted(string $message, string $exceptionMessage, bool $isDebug): self
+    public static function formatted(string $message, string $exceptionMessage, bool $isDebug, ?\Throwable $throwable = null): self
     {
         $formattedMessage = $message;
 
@@ -14,6 +14,6 @@ trait ExceptionMessage
             $formattedMessage .= $exceptionMessage;
         }
 
-        return new self($formattedMessage);
+        return new self(message: $formattedMessage, previous: $throwable);
     }
 }